+++ genium/src/test/resources/smoke/external_types/output/cpp/src/smoke/FooClass.cpp
@@ -11,11 +11,17 @@
 namespace genium {
 }
 namespace smoke {
-static_assert(std::is_same<
-    ::std::string,
-    typename std::result_of<decltype(&::foo::BarClass::doStuff)(::foo::BarClass*, const int32_t)>::type
->::value);static_assert(std::is_same<
-    ::foo::BarEnum,
-    typename std::result_of<decltype(&::foo::BarClass::doStaticStuff)(const ::foo::BarEnum)>::type
->::value);
+static_assert(
+    std::is_same<
+        ::std::string,
+        typename std::result_of<decltype(&::foo::BarClass::doStuff)(::foo::BarClass*, const int32_t)>::type
+    >::value,
+    "Expected '::std::string' return type for '::foo::BarClass::doStuff'."
+);static_assert(
+    std::is_same<
+        ::foo::BarEnum,
+        typename std::result_of<decltype(&::foo::BarClass::doStaticStuff)(const ::foo::BarEnum)>::type
+    >::value,
+    "Expected '::foo::BarEnum' return type for '::foo::BarClass::doStaticStuff'."
+);
 }